Fixed bug where no items are in the root: __getItemFolder__
Fixed logic error in __getItemFolder__ that shows up when there are no
items in the root folder.
Added the option to pass folder ID into each update function. This
increases performance when you already have the folder ID because each
function doesn't need to search for the appropriate folder for the item.